+++ piecrust/src/store/tree.rs
@@ -66,7 +66,7 @@ impl ContractIndex {
         let memory = memory.read();
         let memory_inner = memory.inner;
 
-        element.len = memory_inner.current_len;
+        element.len = memory_inner.mmap.len();
 
         for (dirty_page, _, page_offset) in memory_inner.mmap.dirty_pages() {
             element.offsets.insert(page_offset);